requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.4
idna==2.5
PyYAML==3.12
requests==2.18.1
urllib3==1.21.1
lxml==4.2.1
python-dateutil==2.8.0
